backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pypdf
google-generativeai
python-multipart
fastapi-cors